+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# [5.44.0](https://github.com/biancoroyal/node-red-contrib-modbus/compare/v5.41.0...v5.44.0) (2025-08-19)
+# [6.0.0-beta.1](https://github.com/biancoroyal/node-red-contrib-modbus/compare/v5.41.0...v6.0.0-beta.1) (2025-08-20)
 
 
 ### Bug Fixes
@@ -6,6 +6,7 @@
 * **buffer:** nodejs Buffer alloc update ([9a6350c](https://github.com/biancoroyal/node-red-contrib-modbus/commit/9a6350cd67361f10310629a145f8a595e9f3d85c))
 * **codeql:** version codeql ([163101f](https://github.com/biancoroyal/node-red-contrib-modbus/commit/163101f91c784a5202acd86b0a10302a22414495))
 * **config:** [#442](https://github.com/biancoroyal/node-red-contrib-modbus/issues/442) fixing optional to save correct ([de88e25](https://github.com/biancoroyal/node-red-contrib-modbus/commit/de88e25b10e7e92ea8836d22ee91cb25796f3909))
+* Correct Modbus server buffer addressing for registers ([b95918f](https://github.com/biancoroyal/node-red-contrib-modbus/commit/b95918f9921757e8552d5a5c06b3a3ca531c3b3b)), closes [#537](https://github.com/biancoroyal/node-red-contrib-modbus/issues/537)
 * **issue#472:** Implemented message queue to handle incoming messages asynchronously ([65ea2a7](https://github.com/biancoroyal/node-red-contrib-modbus/commit/65ea2a7dce2e391a0b46412b798fdc4c14e6f750)), closes [issue#472](https://github.com/issue/issues/472)
 * **modbus-client-core:** fixed test cases ([0697df3](https://github.com/biancoroyal/node-red-contrib-modbus/commit/0697df3f4fdf71434e3757f9ae68c6e6db22a948))
 * **modbus-client-flow:** fixed modbus client flow example ([8613c89](https://github.com/biancoroyal/node-red-contrib-modbus/commit/8613c892c54d7c73ca818a5fb34729232fb8acef))